feat: Initial code for GeneralDB
