+++ worker/src/polkadex_db/mod.rs
@@ -1,0 +1,2 @@
+pub mod general_db;
+pub use general_db::*;
